% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_amino_acid}
\alias{is_amino_acid}
\alias{is_peptide}
\title{Assertions}
\usage{
is_amino_acid(x)

is_peptide(x)
}
\arguments{
\item{x}{an object to examine}
}
\value{
logical
}
\description{
Assertions
}
\seealso{
\link[assertthat:assert_that]{assert_that}
}
