% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide.R
\name{peptide}
\alias{peptide}
\alias{as_peptide}
\title{Peptide}
\usage{
peptide(x)

as_peptide(x)
}
\arguments{
\item{x}{a string}
}
\value{
an obect of class \code{peptide}
}
\description{
Simple class for peptidic sequences. Display the sequence in colors in the console. Also works in a \link[tibble:tibble-package]{tibble}.
}
\examples{
peptide("KIAALKE")
}
