% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideview-package.R
\docType{package}
\name{peptideview-package}
\alias{peptideview}
\alias{peptideview-package}
\title{peptideview: Visualisation of peptide sequences}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jean Manguy \email{jean@manguy.eu} (0000-0002-8427-007X)

}
\keyword{internal}
