% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_peptide}
\alias{view_peptide}
\title{View HTML rendering of proteins and peptides sequences}
\usage{
view_peptide(sequence)
}
\arguments{
\item{sequence}{a string}
}
\description{
View HTML rendering of proteins and peptides sequences
}
\examples{
view_peptide("KIAALKE")
}
